move cd to transifex/utils/common.py
